etc/service/getty*: don't use tmux for serial
The idea here is that a serial console is mostly a debugging thing.
It's where dmesg goes, because it's simple and available from early
boot.  For VMs, serial would normally be ingested into logging
infrastructure.  We run a shell there because we might as well, in
case it's useful for debugging, but running a full interactive tmux
there would frustrate its primary purpose, which is logging.

For interactive use, we designate the virtio-console device, as well
as VTs in future.  virtio-console can handle things like being
resized, so it's much better suited to interacting with a VM than an
emulated serial device.