etc/s6-rc: run services from /ext/svc
Right now, we're just running s6-rc services directly from the the
user data partition.  This is not something we should be doing
long-term, because if we can disallow the user data partition from
running arbitrary code, we can have all code executing on the host
system covered by Secure Boot!  But making the user data less dynamic
means we first have to figure out what sort of configuration options,
etc. should be allowed, and we need the freedom to experiment on that
a bit.  Just running arbitrary services provides that freedom for now,
and we can lock it down later once we've discovered what we need to
allow.