host/netvm-vmm: don't set hvc0 as console
I've decided it's better not to do this because console output may
mess with terminal applications the user is running, and the console
output is just a dmesg away.

In default.nix, I've switched back to explicitly symlinking each
thing, because I'd like to be more explicit where each package is
allowed to install files.