etc/service/getty-hvc0: set up terminfo
We have to assume it's foot, because we have no way of knowing for
sure (unless we add a way to communicate it from the host out of
band).

In default.nix, I've gone a bit more for explicitly symlinking each
thing, because I'd like to be more explicit where each package is
allowed to install files.