Discover partitions by UUID
The Spectrum installer is essentially a NixOS system, with a partition
containing a Spectrum live image.  The installer's bootloader also
supports trying out Spectrum by loopback mounting the live image, and
chainloading its bootloader.  But in this case, LoaderDevicePartUUID
is still set to the installer's bootloader, so discovering the rootfs
by looking on the same disk won't work.

Since we're using dm-verity, and know the contents of the filesystem
upfront thanks to the verity hash, we can take advantage of the
Discoverable Partitions Specification[1]'s recommendation to just use
the head of the verity hash as the root partition UUID, and the tail
as the verity partition UUID, and discover them based on that.

In addition to making it possible to try out Spectrum from the
installer, this gives us a nice reduction in complexity.

[1]: https://systemd.io/DISCOVERABLE_PARTITIONS/