host/rootfs: upload core files over VSOCK to host
This is very useful when debugging a crashing program, because the
core file can be loaded into gdb on the host system, which will
hopefully already be set up with the source code of the program,
debugging symbols, etc.

The core dump handler is installed unconditionally such that it'll be
present even if Spectrum isn't running in a VM, or doesn't have a
virtio-vsock device, because there's not much that could usefully be
done with a core file in those situations anyway at the moment.
Eventually we'd probably want to e.g. load it into a VM by default,
when Spectrum is more self-hosting.