host/start-vm: bridge_add: don't accept % in names
Linux will accept %d and dynamically fill it in, but it doesn't update
the string it was given, so there's no way to know what name was
chosen with this API.  If we wanted to support this, we'd probably
have to use the netlink API instead.  Right now, we don't need the
kernel to allocate bridge names for us, so let's just reject the names
we can't handle instead.

We only need to check for the presence of a % character, rather than
doing anything more complicated, because network interfaces aren't
actually allowed to have % in their names, so any % character is
either part of a %d that gets replaced, or will be rejected by the
kernel.