Documentation: enable syntax highlighting
Asciidoctor gives us a bunch of choices for syntax highlighting, so
the choice of Rouge is pretty arbitrary here vs. CodeRay (but I think
it makes more sense than Pygments since Pygments is written in a
different language than Asciidoctor).

Adding an ASCIIDOCTORFLAGS variable that can be overridden with extra
flags seems like a good idea — it was helpful when I tested the
different highlighting options before adding them to the Makefile.