host/rootfs: fix Weston hotplugging
On the Spectrum host, the kernel delivers uevents to mdevd.  By
default, if mdevd doesn't have a handler for an event, it won't do
anything, and other processes that might be interested in the uevent
hear about it.  This means that Weston isn't notified about input
device hotplugs.  (And USB devices can appear after init has started,
so any USB device counts as hotplugging even if it's attached before
boot or even built in to the computer.)

To fix this, we configure mdevd to rebroadcast events it doesn't
handle on Netlink group 4.  These events will then be picked up by
Weston via libudev-zero, which listens on group 4 for hotplug events.

Reported-by: Ville Ilvonen <ville.ilvonen@unikie.com>
Fixes: 372bacd ("etc/s6-rc: add weston service")
Message-Id: <20220424080020.32216-1-hi@alyssa.is>
Tested-by: Ville Ilvonen <ville.ilvonen@unikie.com>