Add fs_freq and fs_passno fields to fstab files
These fields are pretty much obsolete and are supposed to be optional
according to fstab(5), but upstream Musl requires them.  Long ago, I
wrote a patch[1] to fix it, sent it upstream, and then applied it in
Spectrum's Nixpkgs as a temporary measure, to avoid churn in the fstab
files, thinking it would quickly be applied upstream.

But it has now been more than half a year since the last discussion on
my patch.  I do think it'll probably be accepted eventually, but at
this point it's not worth patching Musl ourselves for to avoid the
very minor churn of doing this change now, and reverting it when my
patch is finally accepted.  So just add the fields, and then we can
drop the patch from our Nixpkgs, and then we won't have to rebuild the
Musl-targetting compilers we could otherwise get from Hydra.

[1]: https://inbox.vuxu.org/musl/20210915221155.3977763-4-hi@alyssa.is/