host/start-vm: abort on panic
There's nothing we could do to recover from a panic, and since we use
FFI extensively, to safely use unwinding panics we would have to wrap
every Rust function callable from C in a catch_unwind[1].

It's not currently possible[2] to use this with Rust unit tests, so we
keep panic=unwind there.  It's unlikely that a Rust unit test will
call C which calls Rust anyway.

[1]: https://doc.rust-lang.org/nomicon/unwinding.html
[2]: https://github.com/rust-lang/rust/issues/67650