host/initramfs: use module list from NixOS
This means we don't have to include every kernel module in the
initramfs (making it huge), while we still don't have to maintain our
own list of which ones are required.  The "all-hardware" NixOS profile
is designed for the NixOS installer to be able to boot on as many
systems as possible, which matches pretty well what we need for the
generic Spectrum image.  The makeModulesClosure function from Nixpkgs
would also handle installing firmware in the initramfs, if any of the
modules we were installing required firmware.

An important reason to change this (apart from the 92% size reduction
and the corresponding reduction in boot time) is that, by making every
module available in the initramfs, without including any firmware,
modules that required firmware would be loaded before the firmware
they needed was available.  When the root filesystem became available,
even if it had the firmware on it, it would be too late.