Documentation: reorganize for Diátaxis
The Diátaxis[1] framework promotes separating technical documentation
into Tutorials, How-tos, Reference, and Explanation.  Not all of our
documentation was written to be specifically one of these things, but
each documentation page has one it's closest to, so organising
everything into the proper category is a good start.  It should also
encourage the system to be followed for new documentation.

[1]: https://diataxis.fr/