host/rootfs: disable systemd in weston deps
We're overriding the systemd attribute to point to libudev-zero, so
anything that's trying to build with special systemd support isn't
going to work.  We wouldn't be using that anyway, so just overlay
those packages to disable systemd support.

Long term, we can try to fix this upstream by separating the "udev"
and "systemd" attributes, and only overriding the udev one.  But that
would be quite a bit of work, and making this change here allows us to
drop a bunch of Nixpkgs patches to disable weston dependencies, so
this is still progress.

Signed-off-by: Alyssa Ross <hi@alyssa.is>