Documentation: add a build script
Having two commands that need to be run to build the documentation
isn't very nice, because it means that the only way to confidently
perform a full site build is with nix-build.  And with nix-build, it's
annoying to view the generated site, because there's no analog to
jekyll serve.

So let's extract the build script from Nix, to a shell script.  This
way it can be run by nix-build, or in a nix-shell.  This way, it's
possible to do most work on the site with jekyll serve, and do a full
scripts/build.sh when it's necessary to do things outside of jekyll's
remit, like regenerating the diagram SVGs.

Signed-off-by: Alyssa Ross <hi@alyssa.is>