scripts/make-gpt.sh: reserve 8 MiB at start of GPT
In some cases (e.g. a product built on Spectrum where updates to both
firmware and OS are delivered by the same vendor), it makes sense to
include firmware in the same image as the OS.  This firmware tends to
live in the start of the image, after the GPT header.  (The exact
location varies by board.)

The 1 MiB reserved by sfdisk by default is not enough for this.  The
conventional value seems to be 8 MiB.  On NixOS, this is configurable
with the sdImage.firmwarePartitionOffset option, but this defaults to
8 MiB and I wasn't able to find any examples in searches on DuckDuckGo
and GitHub of anybody using any other value, so it looks like 8 MiB is
broadly acceptable.

Signed-off-by: Alyssa Ross <hi@alyssa.is>