Introduce a build configuration file
By default, a file called "config.nix" in the root of the Spectrum
repository will be read if it exists.  That file should contain an
attribute set.  Currently, only a "pkgs" key is supported, which
allows specifying a custom package set that will be used throughout
the Spectrum Nix files.  This will allow us to provide configuartion
options for people who want to build Spectrum in ways that are
probably not suitable for upstreaming.

For example, using the "pkgs" config option I'm introducing here, it
would be possible to use an overlay to patch individual components,
like so:

	{
	  pkgs = import <nixpkgs> {
	    overlays = [
	      (final: super: {
	        weston = super.weston.overrideAttrs ({ patches ? [], ... }: {
	          patches = patches ++ [
	            path/to/weston.patch
	          ];
	        });
	      })
	    ];
	  };
	}

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-Id: <20220831093727.282797-1-hi@alyssa.is>
Reviewed-by: Ville Ilvonen <ville.ilvonen@unikie.com>
Tested-by: José Pekkarinen <jose.pekkarinen@unikie.com>