host/rootfs: use initramfs in "make run"
This will allow us to stop compiling e.g. the virtio-blk module into
the kernel, because it will be loaded by the initramfs.

This introduces some duplication between the rootfs and initramfs's
Makefiles.  I don't think it's worth the effort at the moment to try
to reduce that, because it would come at the expense of additional
complexity in the Makefiles.  We can revisit this later if we want to.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-Id: <20220901104629.863380-1-hi@alyssa.is>
Tested-by: José Pekkarinen <jose.pekkarinen@unikie.com>