host/rootfs: remove kernel override
These drivers should be loaded by the initramfs if required — most
Spectrum installs won't need the virtio drivers on the host, and
overriding the kernel means more stuff we can't reuse from
cache.nixos.org.

We'll probably want to build in the driver for whatever filesystem we
end up using for the root file system eventually, since it will always
be required, but that should be done as part of a more systematic
effort to optimise our kernel configuration.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-Id: <20220901104629.863380-2-hi@alyssa.is>
Tested-by: José Pekkarinen <jose.pekkarinen@unikie.com>