shell.nix: automatically configure b4
This will allow people who just want to test one patch series to skip
an annoying setup process.  To be honest, I'm not sure if it's worth
keeping the documentation about manually setting it up.  I'm worried
it'll just confuse people.  But I'm keeping it for now.  We can remove
it later.

Signed-off-by: Alyssa Ross <hi@alyssa.is>