Documentation: clarify value of spectrum-config
Before, it was not clear that spectrum-config should be set to a file
path, rather than a directory — i.e. whether it override the root of
the source tree, or the config.nix path within it.  This change
clarifies that spectrum-config should be set to a full
configuration.nix file path, not the directory that contains it.

Reported-by: Ville Ilvonen <ville.ilvonen@unikie.com>
Signed-off-by: Alyssa Ross <alyssa.ross@unikie.com>
Message-Id: <20220921133341.3120867-1-hi@alyssa.is>
Reviewed-by: Ville Ilvonen <ville.ilvonen@unikie.com>