host/rootfs: make the root filesystem read-writable
By default Spectrum OS root filesystem is read only. These changes make
it writable and add 500 MB of free space. During development it is very
useful to have the ability to copy files to root filesystem. It
simplifies debugging and speeds up development by eliminating the need
to rebuild the whole rootfs for every change. This should only be used
for development purposes.

Signed-off-by: Yuri Nesterov <yuriy.nesterov@unikie.com>